EnumClass MonsterType 삭제
구현하다 보니 MonsterType이 move 함수를 갖게 되었다.
이것은 EnumClass인 MonsterType에 맞지 않다고 판단하여 삭제하였다
